move useRefState to hooks directory
